pyproject.toml
[project]
name = "vision-llm-extraction"
version = "0.1.0"
description = "Structured Data Extraction with Vision LLMs and Pydantic - Workshop by learnwithparam.com"
authors = [{ name = "Param Harrison", email = "param@learnwithparam.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "google-generativeai>=0.8.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.6",
]
